create task related to loged in user function added

+++ src/tasks/tasks.service.ts
@@ -1,42 +1,66 @@
+  
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { TaskStatus } from './task-status.enum';
-import { v4 as uuid } from 'uuid';
-import { CreatTaskDto } from './dto/create-task.dto';
+import { CreateTaskDto } from './dto/create-task.dto';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
+import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
+import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
+
 @Injectable()
 export class TasksService {
-    constructor(
-        @InjectRepository(TaskRepository)
-        private taskRepository: TaskRepository
-    ) { }
-    async getAllTasks(): Promise<Task[]> {
-        return this.taskRepository.find();
-    }
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
-        if (!found) throw new NotFoundException
+  constructor(
+    @InjectRepository(TaskRepository)
+    private taskRepository: TaskRepository,
+  ) {}
 
-        return found
+  async getTasks(
+    filterDto: GetTasksFilterDto,
+    user: User,
+  ): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
+  }
 
-    }
+  async getTaskById(
+    id: number,
+    user: User,
+  ): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
-    async creatTask(creatTaskDto: CreatTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(creatTaskDto)
-    }
-
-
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
-        task.status = status;
-        await task.save();
-        return task;
+    if (!found) {
+      throw new NotFoundException(`Task with ID "${id}" not found`);
     }
 
-    async deleteTaskById(id: number): Promise<any> {
-        const result = await this.taskRepository.delete(id);
-        if (result.affected === 0) throw new NotFoundException;
-        return { "msg": `task with id: ${id} deleted` }
+    return found;
+  }
+
+  async createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+  ): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
+  }
+
+  async deleteTask(
+    id: number,
+    user: User,
+  ): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
+
+    if (result.affected === 0) {
+      throw new NotFoundException(`Task with ID "${id}" not found`);
     }
+  }
+
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
+    task.status = status;
+    await task.save();
+    return task;
+  }
 }
